+++ postmarker/tornado.py
@@ -1,5 +1,4 @@
 # coding: utf-8
-from ._compat import get_args
 from .core import PostmarkClient
 
 
@@ -12,15 +11,8 @@ class PostmarkMixin(object):
     def postmark_client(self):
         self.require_setting(POSTMARK_SERVER_TOKEN, 'Postmark client')
         if not hasattr(self, '_postmark_client'):
-            self._postmark_client = PostmarkClient(**self.get_postmark_kwargs())
+            self._postmark_client = PostmarkClient.from_config(self.settings)
         return self._postmark_client
-
-    def get_postmark_kwargs(self):
-        return dict(
-            (arg, self.settings['postmark_' + arg])
-            for arg in get_args(PostmarkClient)
-            if 'postmark_' + arg in self.settings
-        )
 
     def send(self, *args, **kwargs):
         return self.postmark_client.emails.send(*args, **kwargs)